MozReview-Commit-ID: IN2hMCjMHLL

+++ security/certverifier/NSSCertDBTrustDomain.cpp
@@ -30,7 +30,6 @@
 #include "pkix/pkixnss.h"
 #include "prerror.h"
 #include "prmem.h"
-#include "prprf.h"
 #include "secerr.h"
 
 #include "CNNICHashWhitelist.inc"
@@ -438,7 +437,7 @@ NSSCertDBTrustDomain::CheckRevocation(EndEntityOrCA endEntityOrCA,
     // able to fetch a more recent one.
     MOZ_LOG(gCertVerifierLog, LogLevel::Debug,
            ("NSSCertDBTrustDomain: cached OCSP response: error %d",
-           cachedResponseResult));
+            static_cast<int>(cachedResponseResult)));
     // When a good cached response has expired, it is more convenient
     // to convert that to an error code and just deal with
     // cachedResponseResult from here on out.
